// src/app/enums/router-event-type.enum.ts
export enum RouterEventType {
  NAVIGATION = 'NAVIGATION',
  ROUTE_CONFIG_LOAD = 'ROUTE_CONFIG_LOAD',
  ROUTES_RECOGNIZED = 'ROUTES_RECOGNIZED',
  GUARD_CHECK = 'GUARD_CHECK',
  CHILD_ACTIVATION = 'CHILD_ACTIVATION',
  ACTIVATION = 'ACTIVATION',
  RESOLVE = 'RESOLVE',
  SCROLL = 'SCROLL',
  UNHANDLED = ''
}